Consolidate Status > Tables page into a shared web-common component
- Move `LocalProjectTables.svelte` to
  `web-common/src/features/preview-mode/StatusTablesPage.svelte` and
  the supporting query helpers to
  `web-common/src/features/preview-mode/tables-selectors.ts`.
- Both `/status/tables` (local) and `/-/edit/status/tables` (cloud
  editor dev preview) now thin-wrap the shared page.
- Update local `TablesSection` to import the selector from web-common.

The cloud non-edit `/-/status/tables` keeps using web-admin's
`ProjectTables` for now — that surface is out of scope for this PR.
